fixing pycodestyle

+++ tests/test_models/test_base_model.py
@@ -22,10 +22,10 @@ class test_basemodel(unittest.TestCase):
         pass
 
     def tearDown(self):
-        try:
+        """Remove storage file at end of tests"""
+        if (os.path.exists('file.json')):
             os.remove('file.json')
-        except:
-            pass
+        
 
     def test_default(self):
         """ """
